+++ resources/views/design_queue/index.blade.php
@@ -33,6 +33,7 @@
                     <th>Qty</th>
                     <th>Priority</th>
                     <th>Binding Type</th>
+                    <th>Lamination Type</th>
                     <th>Special Note</th>
                     <th>Status</th>
                 </tr>
@@ -66,6 +67,7 @@
 
                     <td><span class="btn btn-sm {{ $priorityClass }}">{{ $designQueue->orderedBook->order->order_priority ?? 'N/A' }}</span></td>
                     <td>{{ $designQueue->orderedBook->binding_type }}</td>
+                    <td>{{ $designQueue->orderedBook->lamination_type }}</td>
                     <td>{{ $designQueue->orderedBook->special_note ?? '-' }}</td>
                     <td>
                         <form action="{{ route('design-queues.update', $designQueue->id) }}" method="POST" class="form-inline">