+++ resources/views/parts/footer.blade.php
@@ -1,5 +1,5 @@
 <footer class="footer-wrap pd-20 mb-20 card-box">
-    CrynoERP - v1.0.1
+    CrynoERP - v1.0.3
 </footer>
 
 <style>